Add files via upload
This code is functioning, but it is a stub and has some issues right now. For example very high magnetic fields h>50 may give some issues for T large. Also the code needs to be cleaned up an commented, a file recapping and elaborating on the talk where this code was presented (student workshop on Integrability, Louvain-la-neuve 2019, https://integrability2019.pj4e.de) will be presented at later times